pyproject.toml
[build-system]
requires-python = '==3.10.*'
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimyria"
version = "0.9.0"
dependencies = [
  'ase',
  'numpy',
  'e3nn==0.5.6',
  'pyzstd',
  'pyyaml'
]

[project.scripts]
mimyria-py = "mimyria.cli:main"

[project.entry-points."ase.ioformats"]
polar = "mimyria.io:PolarIOFormat"

[tool.setuptools]
packages = ["mimyria"]

